moved tillage implement enums into EEE enums
