+++ RUFAS/routines/field/field/tillage_application.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from RUFAS.routines.field.field.field_data import FieldData
 from RUFAS.routines.field.soil.soil_data import SoilData
-from RUFAS.routines.tillage_implements_enum import TillageImplement
+from RUFAS.routines.EEE.enums import TillageImplement
 from RUFAS.output_manager import OutputManager
 
 om = OutputManager()
