+++ tests/soil_crop_tests/field_tests/test_field.py
@@ -34,7 +34,7 @@ from RUFAS.routines.manure.manure_nutrients.nutrient_request_results import (
     NutrientRequestResults,
 )
 from RUFAS.routines.manure.manure_nutrients.nutrient_request import NutrientRequest
-from RUFAS.routines.tillage_implements_enum import TillageImplement
+from RUFAS.routines.EEE.enums import TillageImplement
 
 om = OutputManager()
